+++ client/src/components/Header/SearchBar.tsx
@@ -11,19 +11,27 @@ function SearchBar() {
     if (mapInstance) {
       const geocoder = new MapboxGeocoder({
         // Initialize the geocoder
-        accessToken:
-          'pk.eyJ1IjoiY2hyaXNjZXJpZSIsImEiOiJja3VvcXBiaGExcG5vMnFtYjhnc3gxcGprIn0.eX9g2ClfVBqYEvecwIPLYw', // Set the access token
+        accessToken: process.env.REACT_APP_MAPBOX_ACCESS_TOKEN || '',
         mapboxgl: mapInstance, // Set the mapbox-gl instance
-        marker: false, // Do not use the default marker style
+        marker: true, // Do not use the default marker style
         placeholder: 'Search', // Placeholder text for the search bar
         trackProximity: true,
+        types: 'place, poi, address',
       });
       geocoder.addTo('#mapbox-geocoder-container');
+
+      const setProximity = () => {
+        var center = mapInstance.getCenter().wrap();
+        geocoder.setProximity({ longitude: center.lng, latitude: center.lat });
+      };
+
+      mapInstance.on('load', setProximity);
+      mapInstance.on('moveend', setProximity);
 
       geocoder.on('result', (e) => {
         mapInstance.flyTo({
           center: e.result.geometry.coordinates,
-          zoom: 16,
+          zoom: 17.5,
         });
       });