+++ src/routes/users.ts
@@ -1,4 +1,6 @@
 import express from 'express';
+import { Error } from 'mongoose';
+import User, { IUser } from '../models/usersModel';
 
 const router = express.Router();
 
@@ -8,6 +10,17 @@ router.get('/v1/users/current-user', (req, res) => {
   } else {
     res.status(404).send(null);
   }
+});
+
+// Get user by id
+router.get('/v1/users/:id', (req, res) => {
+  User.findById(req.params.id, (err: Error, user: IUser) => {
+    if (err) {
+      res.status(500).send(err);
+    } else {
+      res.status(200).send(user);
+    }
+  });
 });
 
 export default router;